Implement ImageDuplication visual transformation
This transformation attempts to solve the following problems:
- Basic C-03
- Basic C-04
- Basic C-06
- Basic C-10
- Challenge C-03
- Challenge C-06
- Challenge C-07
- Challenge C-11

Also remove the UnionTransformation for now since it is causing
the agent to answer problem Basic C-11 out of luck, but of
course fails to generalize to the Test problem, so we should
rather skip it.

Another notable change is the ability now for individual
transformations to specify a degree of confidence to override the
strict 90% imposed before. For example, the ImageDuplication
uses 84% since images are hard to replicate exactly, but should be
fairly similar.

This transformation achieved the following results for Problem Set C:

"Basic Problems C": {
    "Incorrect": "0",
    "Skipped": "6",
    "Correct": "6",
    "Set": "Basic Problems C"
}

"Challenge Problems C": {
    "Incorrect": "1",
    "Skipped": "6",
    "Correct": "5",
    "Set": "Challenge Problems C"
}

"Test Problems C": {
    "Incorrect": "1",
    "Skipped": "9",
    "Correct": "2",
    "Set": "Test Problems C"
}

"Ravens Problems C": {
    "Incorrect": "0",
    "Skipped": "10",
    "Correct": "2",
    "Set": "Ravens Problems C"
}

Not really what I expected! The Test problems are all being skipped
regardless of the new transformation. Maybe I need to lower the
confidence a bit more?