Implement 3x3 semantic relationships for Set C:
- A semantic points system based on shapes' filled-ness
- A semantic relationships that looks at shapes' scaling differences

These semantic relationships attempt to solve Basic Problem C-08, C-02.

With these transformations, I was finally able to obtain the minimum
performance requirement: 11/12 for Basic C and 7/12 for Test C!
I also fixed a sneaky bug that I hadn't notice when working on
Project 1 and, guess what? it allowed me to solve all Test problems
for the set B! Of course it won't count for Project 2, but my
overall accuracy will definitely increase!

Following are the complete results:

"Basic Problems C": {
  "Incorrect": "0",
  "Skipped": "1",
  "Correct": "11",
  "Set": "Basic Problems C"
}

"Challenge Problems C": {
  "Incorrect": "2",
  "Skipped": "4",
  "Correct": "6",
  "Set": "Challenge Problems C"
}

"Test Problems C": {
  "Incorrect": "2",
  "Skipped": "3",
  "Correct": "7",
  "Set": "Test Problems C"
}

"Ravens Problems C": {
  "Incorrect": "1",
  "Skipped": "8",
  "Correct": "3",
  "Set": "Ravens Problems C"
}

Ravens Problems C results are pretty bad, but I feel they are related to
the fact that a lot of my solutions need certain shapes to be there.
My next idea to try is to not require the shapes for the ImageDuplication
transformation and use a bounding box as the proxy for the width and height
to use when moving the images.