Add a rotation + union transformation
This transformation attempts to solve Problem Basic C-11 via its
row-wise transformation. We had previously solved it using the
top-down deletion but that only accounted for its column-wise
transformation, and it did not solve its Test counterpart. Maybe
this other way will help with that.

And it did! Using this new transformation my agent was able to go
from 7/12 to 9/12 Test C problems!

The full results are:

"Basic Problems C": {
    "Incorrect": "0",
    "Skipped": "1",
    "Correct": "11",
    "Set": "Basic Problems C"
}

"Challenge Problems C": {
    "Incorrect": "2",
    "Skipped": "4",
    "Correct": "6",
    "Set": "Challenge Problems C"
}

"Test Problems C": {
    "Incorrect": "1",
    "Skipped": "2",
    "Correct": "9",
    "Set": "Test Problems C"
}

I feel like the two problems that are being skipped have to do
with the ImageDuplication transformation; probably because the
shapes are being extracted incorrectly and the problems are
being skipped. It is hard to know what the problem is, and since
I only have 3 submissions left, I think I am going to focus on
solving more Challenge problems. Maybe while doing that it will
also help solve those other Test problems, we'll see!