Implement an InvertedDiagonalUnion semantic relationship
This relationship was able to solve the Basic Problem C to achieve
a perfect score of 12/12 for the Basic C Set! However, it did not
improve the best score for the Test problems of 9/12, but it is
definitely an improvement!

The full results are as follows:

"Basic Problems C": {
  "Incorrect": "0",
  "Skipped": "0",
  "Correct": "12",
  "Set": "Basic Problems C"
}

"Challenge Problems C": {
  "Incorrect": "2",
  "Skipped": "3",
  "Correct": "7",
  "Set": "Challenge Problems C"
}

"Test Problems C": {
  "Incorrect": "2",
  "Skipped": "1",
  "Correct": "9",
  "Set": "Test Problems C"
}

"Ravens Problems C": {
  "Incorrect": "1",
  "Skipped": "7",
  "Correct": "4",
  "Set": "Ravens Problems C"
}