Implement two semantic relationships for Set D
The semantic relationships are:
- FindMissingFrame
- FindAndMergeCommonShapesRowColumn

This commit also adds a new shape to our templates,
the DIAMOND, which was necessary to be able to solve
Basic Problem D-05. In this commit, I also constraint
the ShapeFillPointsSystem relationship to only be
valid when all shapes are the same since it was causing
some false positives in set D. Finally, I increased the
confidence of the NoOp transformation to 0.86 since it
was causing Basic Problem D-05 to be answered incorrectly.
This number seems to be fitted to the Basic set so it
might not generalize that well.

These changes obtained the following results:

* Set D
  * Basic: 6
  * Test: 4
  * Ravens: 1
  * Challenge: 2

* Set E
  * Basic: 7
  * Test: 6
  * Ravens: 6
  * Challenge: 2