Implement las two semantic relationships for set D:
- FindMissingImagePattern
- FindMissingShapeAndCount

Additionally, modify InvertedDiagonalUnion relationship to validate
whether it actually applies to the given problem or not, to avoid
false positives that were happening in Basic Problem D-09. Also,
blacklist Challenge Problem C-08 since it takes forever and it is
not worthy.

These relationships achieved the following results:

* Set D
  * Basic: 11
  * Test: 7
  * Ravens: 2
  * Challenge: 2

* Set E
  * Basic: 7
  * Test: 7
  * Ravens: 6
  * Challenge: 3

My new relationships were able to solve their Test D counterparts
achieving the minimum result of 7/12 for the Test D set and because
I also achieved 11/12 for the Basic D set, this gave me a 15% of the
performance grade. However, one of these relationships was also able
to solve one of Test E problems! Since we don't have access to that
set, I have no idea which problem it was able to solve, but thanks
to that, I have now 7/12 for both the Basic and Test E sets which
means that with this submission I have secured the full 30% of the
performance grade, awesome!